+++ web/src/components/Header/Nav/index.tsx
@@ -19,12 +19,7 @@
 import React from 'react';
 import classnames from 'classnames';
 
-import {
-  getNewPath,
-  getBooksPath,
-  getRepetitionsPath,
-  getDigestsPath
-} from 'web/libs/paths';
+import { getNewPath, getBooksPath } from 'web/libs/paths';
 import { Filters, toSearchObj } from 'jslib/helpers/filters';
 import Item from './Item';
 import styles from './Nav.scss';
@@ -41,8 +36,6 @@ const Nav: React.FunctionComponent<Props> = ({ filters }) => {
       <ul className={classnames('list-unstyled', styles.list)}>
         <Item to={getNewPath(searchObj)} label="New" />
         <Item to={getBooksPath(searchObj)} label="Books" />
-        <Item to={getRepetitionsPath(searchObj)} label="Repetition" />
-        <Item to={getDigestsPath(searchObj)} label="Digests" />
       </ul>
     </nav>
   );