+++ web/src/components/Header/SearchBar/AdvancedPanel/WordsSearch.tsx
@@ -24,9 +24,14 @@ import styles from './AdvancedPanel.scss';
 interface Props {
   words: string;
   setWords: (string) => void;
+  disabled: boolean;
 }
 
-const WordsSearch: React.FunctionComponent<Props> = ({ words, setWords }) => {
+const WordsSearch: React.FunctionComponent<Props> = ({
+  words,
+  setWords,
+  disabled
+}) => {
   return (
     <section className={styles.section}>
       <label htmlFor="has-words" className={styles.label}>
@@ -39,6 +44,7 @@ const WordsSearch: React.FunctionComponent<Props> = ({ words, setWords }) => {
             styles.input
           )}
           value={words}
+          disabled={disabled}
           onChange={e => {
             const val = e.target.value;
             setWords(val);